r5rs-doc/info.rkt: note lack of license metadata (needs SchemeReport)
The license used for the RNRS documents does not yet have an SPDX
license identifier, so we can't add accurate license metadata to
this package yet. For now, add a comment explaining the situation.

Also, add a note about this license in the LICENSE file.

Related to spdx/license-list-XML#1340
Related to racket/racket#3760